requirements.txt
pyyaml
google_spreadsheet
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
python-telegram-bot[socks]
PySocks